requirements.txt
haversine == 2.8.1
plotly == 5.9.0
folium == 0.15.1
pandas == 1.4.4
streamlit == 1.31.1
Pillow==9.2.0
streamlit-folium==0.18.0